Feat(configuration): Add external config file - Add WebDriverConfig.java configuration class - Fix checkContentLengthOfTable test
