+++ src/main/java/idf/artemis/testngspringexample/config/SpringConfig.java
@@ -1,35 +1,11 @@
 package idf.artemis.testngspringexample.config;
 
-import org.openqa.selenium.NoSuchElementException;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
-import org.openqa.selenium.support.ui.FluentWait;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
-
-import java.time.Duration;
 
 @Configuration
 @ComponentScan("idf.artemis.testngspringexample")
 public class SpringConfig {
 
-  @Bean
-  public WebDriver webDrive() {
-    System.setProperty("webdriver.chrome.driver", "/usr/bin/chromedriver");
-    ChromeOptions chromeOptions = new ChromeOptions();
-    chromeOptions.addArguments("disable-infobar");
-    chromeOptions.setHeadless(false);
-    return new ChromeDriver(chromeOptions);
-  }
-
-  @Bean
-  public FluentWait<WebDriver> fluentWait(WebDriver webDriver) {
-    return new FluentWait<>(webDriver)
-        .withTimeout(Duration.ofSeconds(2))
-        .pollingEvery(Duration.ofMillis(250))
-        .ignoring(NoSuchElementException.class);
-  }
 
 }
